Moving models to hub-detect-model. Fixing nuget and gradle

+++ src/main/groovy/com/blackducksoftware/integration/hub/detect/DetectProjectManager.groovy
@@ -38,10 +38,10 @@ import com.blackducksoftware.integration.hub.bdio.simple.model.BdioProject
 import com.blackducksoftware.integration.hub.bdio.simple.model.DependencyNode
 import com.blackducksoftware.integration.hub.bdio.simple.model.SimpleBdioDocument
 import com.blackducksoftware.integration.hub.detect.bomtool.BomTool
-import com.blackducksoftware.integration.hub.detect.bomtool.output.DetectCodeLocation
-import com.blackducksoftware.integration.hub.detect.bomtool.output.DetectProject
 import com.blackducksoftware.integration.hub.detect.hub.HubSignatureScanner
-import com.blackducksoftware.integration.hub.detect.type.BomToolType
+import com.blackducksoftware.integration.hub.detect.model.BomToolType
+import com.blackducksoftware.integration.hub.detect.model.DetectCodeLocation
+import com.blackducksoftware.integration.hub.detect.model.DetectProject
 import com.blackducksoftware.integration.hub.detect.util.DetectFileManager
 import com.blackducksoftware.integration.util.ExcludedIncludedFilter
 import com.blackducksoftware.integration.util.IntegrationEscapeUtil
@@ -100,10 +100,6 @@ class DetectProjectManager {
                     logger.info("${bomToolTypeString} applies given the current configuration.")
                     foundAnyBomTools = true
                     List<DetectCodeLocation> codeLocations = bomTool.extractDetectCodeLocations()
-                    if (!detectProject.projectName && !detectProject.projectVersionName && bomTool.projectName && bomTool.projectVersion) {
-                        detectProject.projectName = bomTool.projectName
-                        detectProject.projectVersionName = bomTool.projectVersion
-                    }
                     if (codeLocations != null && codeLocations.size() > 0) {
                         detectProject.addAllDetectCodeLocations(codeLocations)
                     } else {
@@ -151,14 +147,18 @@ class DetectProjectManager {
             if (detectConfiguration.aggregateBomName) {
                 aggregateBdioDocument.components.addAll(dependencyNodeTransformer.addComponentsGraph(aggregateBdioDocument.project, it.dependencies))
             } else {
-                final SimpleBdioDocument simpleBdioDocument = createSimpleBdioDocument(detectProject, it)
-                final String filename = it.createBdioFilename(integrationEscapeUtil, detectFileManager, detectProject.projectName, detectProject.projectVersionName)
-                final File outputFile = new File(detectConfiguration.getOutputDirectory(), filename)
-                if (outputFile.exists()) {
-                    outputFile.delete()
+                if (it.dependencies) {
+                    final SimpleBdioDocument simpleBdioDocument = createSimpleBdioDocument(detectProject, it)
+                    final String filename = it.createBdioFilename(integrationEscapeUtil, detectFileManager.extractFinalPieceFromPath(it.sourcePath), detectProject.projectName, detectProject.projectVersionName)
+                    final File outputFile = new File(detectConfiguration.getOutputDirectory(), filename)
+                    if (outputFile.exists()) {
+                        outputFile.delete()
+                    }
+                    final File createdBdioFile = writeSimpleBdioDocument(outputFile, simpleBdioDocument)
+                    bdioFiles.add(createdBdioFile)
+                } else {
+                    logger.debug("Could not find any dependencies for code location ${it.sourcePath}")
                 }
-                final File createdBdioFile = writeSimpleBdioDocument(outputFile, simpleBdioDocument)
-                bdioFiles.add(createdBdioFile)
             }
         }
 
@@ -174,7 +174,7 @@ class DetectProjectManager {
     }
 
     private SimpleBdioDocument createSimpleBdioDocument(DetectProject detectProject, DetectCodeLocation detectCodeLocation) {
-        final String codeLocationName = detectProject.getCodeLocationName(detectFileManager, detectCodeLocation.bomToolType, detectCodeLocation.sourcePath, 'Hub Detect Tool')
+        final String codeLocationName = detectProject.getCodeLocationName(detectCodeLocation.bomToolType, detectFileManager.extractFinalPieceFromPath(detectCodeLocation.sourcePath), 'Hub Detect Tool')
         final String projectId = detectCodeLocation.bomToolProjectExternalId.createDataId()
         final BdioExternalIdentifier projectExternalIdentifier = bdioPropertyHelper.createExternalIdentifier(detectCodeLocation.bomToolProjectExternalId)
